refactor: merged order services into one, removed unused helper, added populate method to productsApiService
